+++ testauth/urls.py
@@ -1,7 +1,13 @@
-from django.conf.urls import include, url
+from django.conf.urls import include
+from django.urls import re_path
 
 from allianceauth import urls
 
 urlpatterns = [
-    url(r"", include(urls)),
+    re_path(r"", include(urls)),
 ]
+
+handler500 = "allianceauth.views.Generic500Redirect"
+handler404 = "allianceauth.views.Generic404Redirect"
+handler403 = "allianceauth.views.Generic403Redirect"
+handler400 = "allianceauth.views.Generic400Redirect"